feat: TicketFactory.class 구현 및 수동/자동 로또 상속
