+++ src/main/java/lotto/domain/factory/TicketFactory.java
@@ -1,0 +1,7 @@
+package lotto.domain.factory;
+
+import lotto.domain.LottoTicket;
+
+public interface TicketFactory {
+    LottoTicket create();
+}
